Fix for ignoring fall-through profile data when jump is followed by no-op
Summary:
When a conditional jump is followed by one or more no-ops, the
destination of fall-through branch was recorded as the first no-op in
FuncBranchInfo. However the fall-through basic block after the jump
starts after the no-ops, so the profile data could not match the CFG
and was ignored.

(cherry picked from commit 620273f0f7e1dbd57f82aa28b3c5803ad3a16078)

Differential Revision: https://phabricator.intern.facebook.com/D3496084