Basic block clustering algorithm for minimizing branches.
Summary:
This algorithm is similar to our main clustering algorithm but uses
a different heuristic for selecting edges to become fall-throughs.
The weight of an edge is calculated as the win in branches if we choose
to layout this edge as a fall-through. For example, the edges A -> B with
execution count 100 and A -> C with execution count 500 (where B and C
are the only successors of A) have weights -400 and +400 respectively.

(cherry picked from commit a04e6b48247d2c9e1536e9a0860bb0b2a6c94a12)

Differential Revision: https://phabricator.intern.facebook.com/D3606591