Fix hfsort callgraph stats, add hfsort test.
Summary:
The stats for call sites that are not included in the call graph were broken.
The intention is to count the total number of call sites vs. the number of call sites that are ignored because they have targets that are not BinaryFunctions.

Also add a new test for hfsort.

(cherry picked from commit ff5a11a1ac991a93c4463c11d00c98d24d6816db)

Differential Revision: https://phabricator.intern.facebook.com/D4668631