[BOLT] Fix SCTC issue with hot-cold split
Summary:
SCTC was deleting an unconditional branch to a block in the
cold area because it was the next block in the layout vector. Fix the
condition to only delete such branches when source and target are in
the same allocation area (either both hot or both cold).

(cherry picked from commit 4898696f3ea0fb0994308dccd38d8bd26e8dacfa)

Differential Revision: https://phabricator.intern.facebook.com/D5570300