Fix SCTC bug when two pred/succ BB are in a loop.
Summary: It's possible that two basic blocks being conidered for SCTC are in a loop in the CFG.  In this case a block that is both a predecessor and a successor may have been processed and marked invalid by a previous iteration of the SCTC loop. We should skip rewriting in this case.

(cherry picked from commit c57befbb5f6ad7be8f95f01109c263d587f85b67)

Differential Revision: https://phabricator.intern.facebook.com/D5886721