[BOLT] Ignore Clang LTO artifact file symbol
Summary:
The presence of ld-temp.o symbol is somewhat indeterministic.
I couldn't find out exactly when it's generated, it could be
related to LTO vs ThinLTO, but not always.

If the symbol is there, it could affect names of most
of functions in LTO binary. The status of the symbol
may change between the binary the profile was collected on,
and the binary BOLT is called on. As a result, we may mismatch
many function names.

It is safe to ignore this symbol.

(cherry picked from commit 1f10c82bb4cc07791d8554de40da052a295dead4)

Differential Revision: https://phabricator.intern.facebook.com/D5908955