[BOLT] Always call fixBranches in relocation mode.
Summary:
If you attempted to use a function filter on a binary when in relocation mode, the resulting binary would probably crash.  This is because we weren't calling fixBranches on all functions.  This was breaking bughunter.sh

I also strengthened the validation of basic blocks.  The cond branch should always be non-null when there are two successors.

(cherry picked from commit e2de416fea7d6852af110084d8450261c9292c95)

Differential Revision: https://phabricator.intern.facebook.com/D6261930