Parallelize ICF Pass
Summary:
ICF consumes 10-15% of bolt runtime, for HHVM that is around 45 seconds.
this diff perform some parallelization for the pass to make it faster.
A 60% reduction in the ICF runtime  is measured on the parallel version for HHVM.

(cherry picked from commit 412861f510a1c4495670b0d9a19a034ba3126e77)

Differential Revision: https://phabricator.intern.facebook.com/D15589515