Run hfsort+ in parallel
Summary:
hfsort+ performs an expensive analysis to determine the
new order of the functions. 99% of the time during hfsort+
is spent in the function runPassTwo. This diff runs the body
of the hot loop in runPassTwo in parallel speeding up the
total runtime of reorder-functions pass by up to 4x

(cherry picked from commit f80963655c60f24aa2ffed9872127f2f742330d1)

Differential Revision: https://phabricator.intern.facebook.com/D16450780