Rebase: [lld] second try at --discard-section
Summary:
We don't use .debug_ranges for symbolication but gdb uses it, so we decided to keep the section in D15950051.  However, the section can be quite large (e.g. 543.7 MB in top_aggregator_server).  Even though we can use /DISCARD/ to discard sections in a linker script, it's easier to add a flag so we can use it for Sandcastle builds.
Implementation for gold to follow.

The first attempt (D16009512) was reverted because we cannot discard sections if `--emit-relocs` or `--relocatable` is used (D16198993). Added a check if `config->copyRelocs` is true and added an extra unittest.

This now also contains the work from D25058042 and D44846489.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D16867134
https://phabricator.intern.facebook.com/D23376752
https://phabricator.intern.facebook.com/D26052127
https://phabricator.intern.facebook.com/D33289257
https://phabricator.intern.facebook.com/D38301288
https://phabricator.intern.facebook.com/D41552242

Test Plan: ninja check-lld

Reviewers: #compiler_toolchain

Subscribers: moti

Differential Revision: https://phabricator.intern.facebook.com/D44846367

Tasks: T46459577