[llvm-create-profile] Add switch to include empty binary
Summary: Add "-fill-missing-binaries-with-zero" switch to llvm-create-profile tool so we can include a binary even if there's no sample hit within that binary. With this switch, we can generate an empty (filled with zero counts) AutoFDO profile which preserves inline tree structure. Then we can use a script to parse the profile to construct an inline tree for analysis. The new switch is off be default, so it won't have impact on existing usage. 

Test Plan: Tested with the new script for generating inline log without profile.

Reviewers: twoh, #compiler_toolchain

Reviewed By: twoh

Subscribers: phabricatorlinter

Differential Revision: https://phabricator.intern.facebook.com/D16907939

Signature: 16907939:1566275957:d6f73bea18a7413f3266879b27de6913d74ffe66