Rebase: Block label inserter pass
Summary:
This diff implements block label inserter pass, which inserts machine block label at the beginning of the basic block as a form of inline asm. The label shows up in the binary symbol table with the address, so we can use this information to map the perf2asm result back to MIR block and see how precise the profile data is. The label is a concatenation of function GUID, block ID, and the profile count, so that from the label itself we can tell the profile count of the block.

I tested it with AdIndexer (although the implementation of the pass is slightly changed since then. For example, I initially made the pass to print the auxiliary output with block label and profile count mapping, but later changed the design to encode profile count into the label itself), and was able to find some interesting discrepancies between the profile count and the actual sample count. https://fb.quip.com/eX0VAyvv5CYU has some of these examples, if you're interested in.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D16982244
https://phabricator.intern.facebook.com/D17278104
https://phabricator.intern.facebook.com/D18194355
https://phabricator.intern.facebook.com/D18278030
https://phabricator.intern.facebook.com/D19499182
https://phabricator.intern.facebook.com/D19934199
https://phabricator.intern.facebook.com/D19970004
https://phabricator.intern.facebook.com/D20002182
https://phabricator.intern.facebook.com/D20341821
https://phabricator.intern.facebook.com/D20355506
https://phabricator.intern.facebook.com/D20366189
https://phabricator.intern.facebook.com/D20453348
https://phabricator.intern.facebook.com/D20461731
https://phabricator.intern.facebook.com/D20492044
https://phabricator.intern.facebook.com/D20536321
https://phabricator.intern.facebook.com/D20604094
https://phabricator.intern.facebook.com/D20687736
https://phabricator.intern.facebook.com/D20717544
https://phabricator.intern.facebook.com/D20721658
https://phabricator.intern.facebook.com/D20844118
https://phabricator.intern.facebook.com/D23406197
https://phabricator.intern.facebook.com/D25288113
https://phabricator.intern.facebook.com/D26455796
https://phabricator.intern.facebook.com/D31116883
https://phabricator.intern.facebook.com/D33099387
https://phabricator.intern.facebook.com/D35362902
https://phabricator.intern.facebook.com/D36868401
https://phabricator.intern.facebook.com/D38990229
https://phabricator.intern.facebook.com/D41808425
https://phabricator.intern.facebook.com/D42747283
https://phabricator.intern.facebook.com/D45237625

Test Plan: unit test, AdIndexer.

Reviewers: #compiler_toolchain

Subscribers: wenlei, phabricatorlinter

Differential Revision: https://phabricator.intern.facebook.com/D47127631

Tasks: T48837209