Rebase: [AutoFDO] Tune down compile time opts for better LTO time profile accuracy
Summary:
This is part of the changes that got us ~1% win for mysqld with MonoLTO+AutoFDO initially. For mysqld, the original changes I have were for legacy pass manager, but I ended up using ThinLTO+NewPM because these change were already done by upstream for that pipeline. This one is for MonoLTO+NewPM - it is essentially applying the changes that preserve profile quality for LTO time that are exclusively for ThinLTO+NewPM, to MonoLTO+NewPM as well. Specifically, that's tuning down hot inlining and loop unroll for compile time passes.

With this change, we may be able to get better perf for MonoLTO by enabling NewPM just like what we did for ThinLTO (we could even apply similar changes to MonoLTO+LegacyPM). But I'm doing this not for improving MonoLTO, instead this is for setting up a good MonoLTO base for experimenting proper cross module adjustment of profile based on inline decision which ThinLTO isn't capable of doing right now.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D17708866
https://phabricator.intern.facebook.com/D19499194
https://phabricator.intern.facebook.com/D21241660
https://phabricator.intern.facebook.com/D21530854
https://phabricator.intern.facebook.com/D21587607
https://phabricator.intern.facebook.com/D21803580
https://phabricator.intern.facebook.com/D23376809
https://phabricator.intern.facebook.com/D24657067
https://phabricator.intern.facebook.com/D25234579
https://phabricator.intern.facebook.com/D25907614
https://phabricator.intern.facebook.com/D25929142
https://phabricator.intern.facebook.com/D25942201
https://phabricator.intern.facebook.com/D30978253
https://phabricator.intern.facebook.com/D35903599

Test Plan: mysqld build; ninja check

Reviewers: #compiler_toolchain

Subscribers: marksan, rafaelauler, phabricatorlinter, mren, maks

Differential Revision: https://phabricator.intern.facebook.com/D45539927

Tasks: T54911391