[perf2bolt] Better mmap event matching
Summary:
When perf tool reports a mapping address of a binary, it is not always
the address of the first loadable segment we were checking against.
As a result, perf2botl was not working properly for binaries where the
first segment was not executable.

The fix is to check if the address reported by mmap event matches any
of the loadable segments. Note that the segment alignment has to be
applied to get real loadable address of the segment.

Fixes #65

(cherry picked from commit e132b254dc2356b73fb42e3705b29a6334067121)

Differential Revision: https://phabricator.intern.facebook.com/D19146419