Rebase: Hook up the -fpersist-block-annotation switch with lld
Summary:
Currently the switch only works with the gold linker plug-in. It should be hooked up to lld as well so that we can start using the feature with lld.

Also fixing a test issue where the immediate lL file was left over in place after a test run.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D19669981
https://phabricator.intern.facebook.com/D19767989
https://phabricator.intern.facebook.com/D20826714
https://phabricator.intern.facebook.com/D20894870
https://phabricator.intern.facebook.com/D21481094
https://phabricator.intern.facebook.com/D22877900
https://phabricator.intern.facebook.com/D25175247
https://phabricator.intern.facebook.com/D25256032
https://phabricator.intern.facebook.com/D26028182
https://phabricator.intern.facebook.com/D31116943

Test Plan:
check-clang
check-llvm

Reviewers: #compiler_toolchain

Subscribers: wenlei, twoh, phabricatorlinter, christylee

Differential Revision: https://phabricator.intern.facebook.com/D34471015

Tags: bootcamp