Rebase: [CUDA] add option to allow host device functions to call host functions
Summary:
In CUDA, it's an error to call a host function from a device function or the other way around. In some circumstances, in the device pass compilation (building for GPU), it is allowed to call a host function from a host-device function, as long as the call is not executed: so-called wrong-side rule. NVCC provides a warning at build time (ignored for system headers) and generates a hard error if such a call happens at runtime. Clang also makes it a runtime error but also tries to diagnose it at compile time making it a hard syntax error. The problem is that Clang does the analysis at the AST level, and sometimes rejects presumably valid code:
```
__host__ void h() {}
__host__ __device__ void hd() {
  if (0) h(); // clang generates an error
}
```
godbolt: https://godbolt.org/z/Wttp3F
Unfortunately, this pattern in commonly used in thrust CUDA library, blocking Clang adoption for CUDA code in fbcode.

This change adds an option to lift these diagnostics from unsurpassable hard syntax error to a warning. This change is a retrofitted of https://reviews.llvm.org/D7841, which was reverted http://reviews.llvm.org/D18416.

**Short CUDA intro**
CUDA extends C with the following function attributes:
`__global__` - aka kernel, the function runs on the GPU, called from the CPU or the GPU.
`__device__` - the function runs on the GPU, called from the GPU.
`__host__` - the function runs on the CPU, called from the CPU, equivalent to a vanilla C functions. It exists primarily to mark device functions as both host:
```
__host__ __device__ int min(int x, int y) { return x < y ? x : y; }
```

Similar to NVCC, Clang compiles the same CUDA code multiple times: once for CPU, and once per GPU architecture. In each compilation it defines a set of macros that can be used to identify the pass and possibly customize behavior of functions:
```
__host__ __device__ int foo() {
  // on host
  // on device
}
```
The problem happens when a host-device function in some cases launches a kernel (calls global-function):
```lang=c++
__device__ void d() {}

__global__ void g() { d(); }

__host__ void h() {
  g<<<1,1,1>>>();
}

__host__ __device__ void hd() {
  // on host
 h();
  // on host
  d();
}
```
First a device compilation pass happens, after preprocessor run the code looks like this:
```
__device__ void d() {}

__global__ void g() { d(); }

__host__ __device__ void hd() {
  // on device
  d();
}
```
Compiler notices that kernel `g` is never called and doesn't generate GPU code for it.

Then a host compilation pass happens, after preprocessor run the code looks like this:
```
__device__ void d() {}

__global__ void g() { d(); }

__host__ void h() {
  g<<<1,1,1>>>();
}

__host__ __device__ void hd() {
  // on host
  h();
}
```
It parses all of the functions, notices that hd function launches kernel `g`, then generates host side code that lunches the kernel.
Then at runtime when hd is called on CPU (aka host) it tries to lunch kernel `g`, and generates an error as there is no corresponding GPU code (aka device) for it. This is what caused: S196313.

Thurst CUDA library [1], provides various host-device function that behave in this fashion. They use the following pattern to workaround this problem in NVCC:
```
__device__ void d() {}

__global__ void g() { d(); }

__host__ void h() {
  g<<<1,1,1>>>();
}

__host__ __device__ void hd() {
  if (!__CUDA_ARCH__) {
    // on host
    h();
  }
  else {
    // on device
    d();
  }
}
```
Here an `if`-statement with predicate value known at compile time guarantees that a host function wouldn't be called from the device. At the same time it marks kernel `g` as used (on AST level) and makes the compiler emit it's GPU code. Unfortunately Clang reject the above code because it doesn't evaluate the predicate on the AST level. The aim of this change is to allow such syntax. My plan is to enable it as warning-as-error in fbcode, so that thrust code (treated as system) would be accepted, and fbcode would comply with clang to be compatible with OSS behavior.

---
Looks like this is a known problem in Clang/CUDA communities [2]. This change likely wouldn't be accepted upstream, and prompt changes in the Thrust library. I'm going to suggest thrust maintainer to make their code compatible with Clang in a way how it was done in the HIP fork of Thrust [3]. Specifically, compare [HIPs parallel_for.h](https://github.com/ROCmSoftwarePlatform/rocThrust/blob/develop/thrust/system/hip/detail/parallel_for.h#L110) with [Thrusts](https://github.com/thrust/thrust/blob/master/thrust/system/cuda/detail/parallel_for.h#L152).

1. https://github.com/thrust/thrust
2. http://lists.llvm.org/pipermail/cfe-dev/2018-November/060070.html
3. https://github.com/ROCmSoftwarePlatform/rocThrust

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D20151498
https://phabricator.intern.facebook.com/D23906297
https://phabricator.intern.facebook.com/D23915266
https://phabricator.intern.facebook.com/D24427595
https://phabricator.intern.facebook.com/D25277460
https://phabricator.intern.facebook.com/D25592248
https://phabricator.intern.facebook.com/D25721444
https://phabricator.intern.facebook.com/D25804830This was missed in the original change.
https://phabricator.intern.facebook.com/D25925165
https://phabricator.intern.facebook.com/D27490837Manual rebase conflict history:
https://phabricator.intern.facebook.com/D20151498
https://phabricator.intern.facebook.com/D23906297
https://phabricator.intern.facebook.com/D23915266
https://phabricator.intern.facebook.com/D24427595
https://phabricator.intern.facebook.com/D25277460
https://phabricator.intern.facebook.com/D25592248
https://phabricator.intern.facebook.com/D25721444
https://phabricator.intern.facebook.com/D25804830This was missed in the original change.
https://phabricator.intern.facebook.com/D25925165
https://phabricator.intern.facebook.com/D27490837

Test Plan:
1. Verify the added tests pass:
```
ninja check-clang
```
2. Verify that build in tp2 succeeds:
```
tp2_build
```
3. Verify that CUDA code in dper can be built with Clang:
```
buck build @mode/opt -c fbcode.cuda_use_clang=true -c cxx.untracked_headers=ignore //fblearner/flow/projects/dper:workflow
```

ninja

Reviewers: #compiler_toolchain

Subscribers: xdwang

Differential Revision: https://phabricator.intern.facebook.com/D29002368

Tasks: T63033708, T39310686, T39311125