[CFGChangeLog] Make order of nodes deterministic
Summary:
The changed test `cfg-change-log.ll` fails in llvm-fb with error:
```
FAIL: LLVM :: CodeGen/X86/cfg-change-log.ll (18492 of 36371)
******************** TEST 'LLVM :: CodeGen/X86/cfg-change-log.ll' FAILED ********************
Script:
--
: 'RUN: at line 2';   /home/zhuhan/server-llvm-build/toolchain-server-master/bin/llc < /data/users/zhuhan/server-llvm/llvm-project/llvm/test/CodeGen/X86/cfg-change-log.ll -O3 -cfg-change-log-funcs=test -o /dev/null 2>&1 | /home/zhuhan/ser
ver-llvm-build/toolchain-server-master/bin/FileCheck /data/users/zhuhan/server-llvm/llvm-project/llvm/test/CodeGen/X86/cfg-change-log.ll
--
Exit Code: 1

Command Output (stderr):
--
/data/users/zhuhan/server-llvm/llvm-project/llvm/test/CodeGen/X86/cfg-change-log.ll:7:15: error: CHECK-NEXT: is not on the line after the previous match
; CHECK-NEXT: CFGChangeLog: changed %bb.{{[0-9]+}}.[[NAME2:[a-zA-Z\.]+]]
              ^
<stdin>:187:2: note: 'next' match was here
 CFGChangeLog: changed %bb.3.loop.backedge(0x56565f8, prof N/A), outgoings 0x5656380:2147483648(100.00%)
 ^
<stdin>:185:103: note: previous match ended here
 -> %bb.0.entry(0x5656278, prof N/A), outgoings 0x56564f0:7158279(0.33%) 0x5656700:2140325369(99.67%)
                                                                                                      ^
<stdin>:186:1: note: non-matching line after previous match is here
 CFGChangeLog: deleted %bb.1.loop(0x5656380, prof N/A), outgoings 0x56564f0:7158279(0.33%) 0x5656700:2140325369(99.67%)
^
```
The test is trying to check that the log shows changing two basic blocks and deleting another, in the exact order of two changes and one deletion. Looking at the test failure, the changes match, but the order doesn't. This is because the `CFG` class stores the nodes in an `unordered_set`.  This makes the order of the nodes indeterministic and makes the test unstable.

Changing the nodes container to be a `std::set`. One tradeoff is now insertion and lookup are of logarithmic complexity compared to constant before. But consider this feature is only for debugging purpose this should be okay.

Will port this to `llvm-fb` with another diff.

Test Plan:
ninja
ninja check
Port change over to llvm-fb and check the change log order is the same as in server-llvm.

Reviewers: hoy, #compiler_toolchain

Reviewed By: hoy

Subscribers: wenlei, phabricatorlinter, christylee, hoy

Differential Revision: https://phabricator.intern.facebook.com/D20263912

Tasks: T47571952

Tags: llvm

Signature: 20263912:1583366616:327f6274d6e33f2b77985f6568305ea47334ca11