[llvm-create-profile] Skip branches between split ranges
Summary:
Branch instructions with the source and target in different split ranges of the same function are accidentally treated as a function call. This may result in

1. A misleading call site annotation in the final merged profile
2. An inaccurate function entry count due to the merge of two split ranges, with the entry count of one range is calculated from cross-range branches.  

These are not an issue when post-bolt sampling sees the same application behavior with pre-bolt sampling, which means cold code is consistently cold pre- and post-bolting. However this may not be true.

In this change, branches crossing regions of the same function are ignored in order to fix the two issues. Cross-function branches are still processed. The quality of function profiles should be consistent with the non-splitting case.




Test Plan:
Ad testing with autofdo refresh from bolted binary looks OK, as expected, where counts of jumping from hot code to cold code doesn't matter much: https://our.intern.facebook.com/intern/experiment_store/experiment/3298539087471/#commit1-commit2

I also tested the change with a local program that was designed to have different behaviors between sampling a bolted binary and a non-bolted binary. This way a Bolt-regarded cold region then became hot during the post-bolt sampling and the profile generated out of it looked good.


Reviewers: wenlei, kaderkeita, #compiler_toolchain

Reviewed By: wenlei

Differential Revision: https://phabricator.intern.facebook.com/D21231869

Tasks: T58930588

Signature: 21231869:1587762584:b56c82e132828e1875286b69794869876bf27959