Rebase: [lld] Ignore --threads for compatibility with old linkers
Summary:
Upstream diff https://reviews.llvm.org/D76885 changed the syntax of the `--threads` switch from being a simple flag to a jointed option with a value. This breaks the downwards compatibility and most fbcode services are not buildable. I'm working this around by separating `--threads` from `--threads=N` and having `--threads` ignored which means lld always run in peak multi-threading mode unless `-threads=N` is specified.

The change may have an effect on existing jobs that don't come with `--threads` in their build mode, which means they don't want a multi-threaded build originally but now they build with maximum number of threads.  I hope this is rare.

The change shouldn't have an effect on how `--threads=N` is consumed with the latest compiler except for `--threads` is now allowed and ignored silently without any symptom. As said, this is a short-term workaround. The ultimate fix should be in the build system (T66511289) and once it's done this change should be reverted.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D21390021

Test Plan: ninja check-lld

Reviewers: #compiler_toolchain

Subscribers: wenlei

Differential Revision: https://phabricator.intern.facebook.com/D32323990

Tasks: T37628453