Rebase: [AutoFDO] add profile density computation to llvm-create-profile
Summary:
AutoFDO performance is sensitive to *profile density*, i.e., the amount of samples in the profile relative to the program size, because profiles with insufficient samples could be inaccurate due to statistical noise and thus hurt AutoFDO performance. A previous investigation showed that AutoFDO performed better on MySQL with increased amount of samples. Therefore, we implement a profile-density computation feature in `llvm-create-profile` to give hints about profile density to users and the compiler.

We define the density of a profile `Prof` as follows:
* For each function `A` in the profile, `density(A) = total_samples(A) / sizeof(A)`.
* `density(Prof) = min(density(A))` for all functions `A` that are *warm* (defined below).

We introduce two knobs for profile density computation:
* *density-function cutoff* `N`: a function is considered *warm* if its total-samples is within top `N` percent of the profile.
* *density threshold* `M`: a profile is good if its profile densities is above `M`.

The two knobs are implemented as options of `llvm-create-profile` now. We are going to tune these knobs with benchmark and service profiles and replace the default values of these options with tuned ones in the next version. (Notes on profile density tuning plan: https://fb.quip.com/anSTAJ7qIkIB.)

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D22521488
https://phabricator.intern.facebook.com/D24923275

Test Plan: ninja check-llvm

Reviewers: #compiler_toolchain

Subscribers: wenlei, wangwei

Differential Revision: https://phabricator.intern.facebook.com/D25933807

Tasks: T43879338