[BOLT] Fix sign issue when validating X86 relocations
Summary:
In analyzeRelocations, we extract the result of the relocation
from binary code to recreate the target of it in a few special cases.
For R_X86_64_32S relocations, however, we were neglecting the
possibility of the encoded value in the instruction to be negative.

(cherry picked from commit 62221b4a987e673e0d63b892f8e88f542a317bbc)

Differential Revision: https://phabricator.intern.facebook.com/D24096347