Rebase: getting rid of isolated flow components
Summary:
A post-processing adjustment of control flow:

The object is responsible for adjusting control flow in a function so as to remove all "isolated" components with positive flow that are unreachable from the entry block. For every such component, we find the shortest path from the entry to an exit passing through the component, and increase the flow by one unit along the path.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D24144379
https://phabricator.intern.facebook.com/D24658495

Test Plan:
spec2006 (O3/O0, no LTO); legend:
- propagateWeights is the default (no-mcf) propagation
- mcf-no-island is the existing mcf (D24120787)
- mcf-island is the newly proposed one
- w+1/w ? w : 1/w are transformations of weights applied in line 2416 of SampleProfile.cpp

{F338794897}

Reviewers: #compiler_toolchain

Subscribers: phabricatorlinter, julianmestre, vliaghat

Differential Revision: https://phabricator.intern.facebook.com/D24685142

Tasks: T68973288