[llvm-create-profile] Do not merge coroutine resume function profile into ramp function
Summary:
For a call into a coroutine resume function, we don't want the stats to be merged into the profile of the ramp function.
To do so, we also load the symbols from the symbol table at the beginning alongside the one from debug info, and latter check the name
from both table for a cross comparison. If the name from symbol table is that from debug table plus ".resume", this is a coroutine resume
 function and we skip the profile accumulation.

Test Plan: Added testcase

Reviewers: wenlei, hoy, #compiler_toolchain

Reviewed By: wenlei, hoy

Subscribers: wenlei, hoy, wlei

Differential Revision: https://phabricator.intern.facebook.com/D25981646

Tasks: T83110851

Signature: 25981646:1611255710:61fabfcfec01852d45e5bb7f787f3bcd1bd74966