Rebase: [CSSPGO] keeping entry count in MCF
Summary:
This diff makes MCF to keep the provided entry count, if it is extracted from a non-dangling probe. It is still possible that MCF will modify the count (e.g., when it is needed to reconnected "islands"). For this case, we sync the entry count computed by MCF and one extracted from profile.
As a result, BFI inference is (almost) a no-op for the counts and match the results of MCF (unless there are irreducible loops).

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D26678516

Test Plan: Ran SPEC17 with O3O3, saw a small improvement in the discrepancy metric

Reviewers: #compiler_toolchain

Subscribers: phabricatorlinter, julianmestre

Differential Revision: https://phabricator.intern.facebook.com/D30268575

Tasks: T68973288