Rebase: [CSSPGO] extending rebalancing of flow among dangling probes
Summary:
Extending rebalancing of flow among dangling blocks (that previously worked for diamonds & triangles):
A "dangling construct" is a subgraph of CFG such that:
- it starts at a non-dangling block, Src, ends at a non-dangling block, Dst;
- all directed paths between Src and Dst go through dangling blocks;
- there are no other jumps (i) outgoing from Src, (ii) incoming to Dst, or (iii) to/from the dangling blocks;
- the dangling blocks form a directed _acyclic_ subgraph (that is, there are no dangling loops)

If we have such a construction, we may easily replace all the dangling blocks with a single edge from Src to Dst, and apply MCF on a reduced instance. Alternatively, we may apply MCF on the original instance and then re-balance the flow via a simple propagation.
I chose the latter approach to keep the implementation simple. Performance-wise the two approaches are the same, as we don't have many dangling blocks.

Manual rebase conflict history:
https://phabricator.intern.facebook.com/D27098271
https://phabricator.intern.facebook.com/D29210310
https://phabricator.intern.facebook.com/D29796317

Test Plan:
Added tests to verify the new logic

Following instructions in T86644461, verified that the flow distribution as expected

Reviewers: #compiler_toolchain

Subscribers: phabricatorlinter, spupyrev

Differential Revision: https://phabricator.intern.facebook.com/D30268607

Tasks: T79499639, T68973288