[llvm-create-profile] Fix address resolution for perf profile
Summary:
llvm-create-profile has a mode that takes perf script as input. In this mode, it process mmap events to get actual load address of text section, and then use that and the preferred load address to compute virtual address for symbolization. However, when getting preferred load address it made an incorrect assumption that the first loadable segment is .text. This was true for gold built binaries, but not so for lld built binaries.

This caused the tool to generate bogus AutoFDO profile for lld built binary using perf script as input. The change follows how this is handled in llvm-profgen to get the preferred load address of first text section.

The problem was caught when trying to move AutoFDO profile collection away from perf2autofdo.py so profiling can run on TWTF hosts without fbcode. This is the same issue that hit IG in the past (T74226536).

Test Plan:
ninja (not spending time for new test case as llvm-create-profile will be replaced by llvm-profgen this year)

verified that spec2017 autofdo profile generated using llvm-create-profile with perf script input is now good with the fix.

Reviewers: hoy, #compiler_toolchain

Reviewed By: hoy

Subscribers: phabricatorlinter

Differential Revision: https://phabricator.intern.facebook.com/D28210135

Tasks: T74226536

Signature: 28210135:1620177030:69eac683535f38d90f34aabffec5aa5df29ba447