fixup! Rebase: Merge BOLT codebase in monorepo
Summary:
Remove special DWARF expressions handling from LLVM.
This is aimed at reducing our changes to LLVM. We originally
added support to LLVM MC layer to be able to encode DWARF expressions
out of an intermediate representation in memory that is based on
DWARFDebugFrame parsing of these expressions. However, this is an
overkill and BOLT does not need the ability to re-encode DWARF
expressions. So, for now, let's remove this support and make BOLT read
DWARF expressions as a blob of data, and write it back to the binary
using cfi_escape, the same method that some LLVM backends (AArch64)
use to encode DWARF expressions into CFIs when they need them.

We are still left with a one line change to LLVM libraries, to allow
us to read the raw bytes of the DWARF expression out of
DWARFDebugFrame parsing structures, so we can write them back with
cfi_escape. But that change will be easier to push upstream.

Test Plan: NFC checks, manually verified CFIs for test case exception-cfi-exprs

Reviewers: maks, #llvm-bolt

Reviewed By: maks

Subscribers: phabricatorlinter

Differential Revision: https://phabricator.intern.facebook.com/D30060991

Tasks: T85949345

Tags: aarch64

Signature: 30060991:1629910956:1ba3e8e908f69c9fd889517547437f64eee6015e