[CSSPGO] profi modifications
Summary:
This diff contains the following improvements for **profi**:
- speeding up min-cost max flow computation by ~4x via early-stop heuristics;
- parameter tuning on using instrumentation-based profile;
- spell-checking all comments;

Test Plan:
Counts quality improvements (O3/MonoLTO):

**SPEC17 (as of 07/29):**

{F636982245}

**Cinder (as of 07/29):**

{F637085655}

We've also tested if the new parameters (in particular, for the entry block) may result in zero-flow for the entire function. To this end, we look at so-called zero-count errors that occur when an inferred block count is zero but its value in the ground-truth is non-zero. Formally, coverage as the percentage of blocks with non-zero frequency in the ground-truth that also have non-zero values in the inferred profile. In addition, we measure accuracy defined as the percentage of blocks with non-zero inferred counts that also have non-zero values in the ground-truth. Here are the two metrics (the higher, the better) on SPEC17 wrt instrumented counts:
NEW:
  coverage = 0.9921 (90604 out of 91330), accuracy = 0.9189 (90604 out of 98605)
OLD:
  coverage = 0.9918 (90581 out of 91330), accuracy = 0.9273 (90581 out of 97683)


Reviewers: wenlei, julianmestre, hoy, wlei

Reviewed By: wenlei, hoy

Differential Revision: https://phabricator.intern.facebook.com/D29737064

Tasks: T83510092, T68973288

Signature: 29737064:1627670468:40b50e69065022456845e9c695c792ba32d4b6dc