fixup! Rebase: Merge BOLT codebase in monorepo
Summary:
Remove RefHandler modifications to DebugInfo lib.
RefHandler is a pretty BOLT-specific modification to DWARF
DebugInfo parsing library that allows us to patch references in
.eh_frame. Remove that from LLVM and move that code to BOLT's side,
making it specific to patching addresses in .eh_frame that need to be
fixed in the case that we are changing the code location but we do not
intend to rewrite .eh_frame completely.

This reduces our changes to LLVM libraries and make our codebase more
compatible with the LLVM ecosystem.

Test Plan: NFC tests

Reviewers: maks, #llvm-bolt

Reviewed By: maks

Subscribers: phabricatorlinter

Differential Revision: https://phabricator.intern.facebook.com/D30465014

Tasks: T85949345

Signature: 30465014:1629911816:c2ad10ae366341642e973ccb8f5fa385674f04b1