[BOLT] Improve cmake configs for opensource
Summary:
Change cmake config in BOLT to only support Linux. In other
platforms, we print a warning that we won't build BOLT.  Change
configs to determine whether we will build BOLT runtime libs. This
only happens in x86 hosts. If true, we will build the runtime and
enable bolt-runtime tests. New tests that depend on the bolt_rt lib
needs to be marked REQUIRES:bolt-runtime. I updated the relevant
tests.  Fix cmake to do not crash when building llvm with a target
that BOLT does not support.

Test Plan:
Build BOLT on Windows (VSCode) and verify that the we
refuse to enable the project with a warning. Build BOLT with all
targets enabled and verify that cmake succeeds.

Reviewers: maks, #llvm-bolt

Reviewed By: maks

Differential Revision: https://phabricator.intern.facebook.com/D31935760

Signature: 31935760:1635285144:15dd12619ec98f9c38473440bffd9779927c6142