[BOLT] Fix crash when trying to resolve external symbols for runtime libs
Summary:
As pointed out by Vladislav in issue 217, if our RTDyld-based
linker fails to locate a symbol, it will crash with segfault. Fix that.

Test Plan:
Removed a symbol inserted by the instrumentation pass and
repro'ed the crash due to the missing symbol. With this patch, BOLT
prints a message indicating to the user that the symbol is missing and
exits without crashing.

Reviewers: aaupov, maks, #llvm-bolt

Reviewed By: aaupov, maks

Subscribers: maks

Differential Revision: https://phabricator.intern.facebook.com/D32481543

Signature: 32481543:1637807060:133005141c47c1695c7005bea6159e7a1213ed0a