[BOLT] Fix profile and tests for nop-removal pass
Summary:
Since nops are now removed in a separate pass, the profile is consumed
on a CFG with nops. If previously a profile was generated without nops,
the offsets in the profile could be different if branches included nops
either as a source or a destination.

This diff adjust offsets to make the profile reading backwards
compatible.

Test Plan: sandcastle

Reviewers: aaupov, #llvm-bolt

Reviewed By: aaupov

Subscribers: phabricatorlinter

Differential Revision: https://phabricator.intern.facebook.com/D33231254

Signature: 33231254:1640052601:515719f99bdefe34ecedae72f96edc8be7627f16