[DWARF] Fix split dwarf debug inlining in mix mode situation.
Summary:
With ThinLTO enabled we can have a function in a CU that doesn't have split
dwarf inlining enabled that gets inlined into CU that has it enabled.
When we try to lookup Abstract Scope Die for the inlined scope in
constructInlinedScopeDIE, it's not present and we assert.

Doing downstream commit until I can get it or better implementation through upstream.
https://reviews.llvm.org/D124892
@opt-out-review

Test Plan: ninja check all

Reviewers: wenlei, #compiler_toolchain

Reviewed By: wenlei

Differential Revision: https://phabricator.intern.facebook.com/D36391434

Tasks: T116556788