AArch64: Add option to keep frame-record (save LR+FP) on top of CSR area
Summary: This adds an LLVM internal `aarch64-frame-record-top` switch to ensure the LR and FP (called frame-record) come first in the CSR save area. https://reviews.llvm.org/D65653 changed LLVM defaults to put the frame record last. With this it is no longer possible to compute the end of the stackframe of the calling function. However some frame unwinding code in HHVM relies on this. This adds a flag to restore the previous behavior.

Test Plan: Added a simple unit test; Inspected assembly output of simple functions. HHVM team reported that this fixes a couple unit tests for their JIT.

Reviewers: wenlei, #compiler_toolchain

Reviewed By: wenlei

Subscribers: ottoni, michaelofarrell, hoy

Differential Revision: https://phabricator.intern.facebook.com/D39635623

Tasks: T130465181

Tags: aarch64