[BOLT][DWARF] Fix for .debug_line with DWARF5
Summary:
There was a bug in a code that pre-populated line string for a case where parts
of .debug_line are not processed by BOLT, but copied as raw data. We were not
switching sections. This resulted in parts of the binary being over-written with
debug data.

Test Plan: ninja-check all, manual check on clang-17

Reviewers: maks

Reviewed By: maks

Differential Revision: https://phabricator.intern.facebook.com/D47164589

Tasks: T154829366