[BOLT][DWARF] Update handling of size 1 ranges and fix sub-programs with ranges
Summary:
When output range is only one entry, and input is low_pc/high_pc do not convert
to ranges. This helps with size of .debug_ranges/.debug_rnglists. It also helps
when either low_pc/high_pc is 0. We not generating potentially invalid ranges
that result in LLDB error.

Also fixed handling of DW_AT_subprogram with ranges. This can be created with
-fbasic-block-sections=all.

Test Plan: ninja check-bolt

Reviewers: rafaelauler, maks, #llvm-bolt

Reviewed By: maks

Subscribers: jianyuhuang

Differential Revision: https://phabricator.intern.facebook.com/D47422946

Tasks: T146873453