[BOLT] Fix handling of trailing entries in jump tables
Summary:
If a jump table has entries at the end that are a result of
__builtin_unreachable() targets, BOLT can confuse them with function
pointers. In this case, we should exclude these targets from the table
as we risk incorrectly updating the function pointers. It is safe to
exclude them as taking such targets is considered an undefined behavior.

Test Plan: Problematic function in unicorn binary after the fix: P1202443323. Before the fix: P1202414748.

Reviewers: aaupov, #llvm-bolt

Reviewed By: aaupov

Differential Revision: https://phabricator.intern.facebook.com/D55552751