[server-llvm-19][AArch64] Replace AND with LSL#2 for LDR target (#34101) (#89531)
Summary:
This backports https://github.com/llvm/llvm-project/commit/77fccb35ac08f66d52bb152735e27572bf9f3f93

Currently, process of replacing bitwise operations consisting of
`LSR`/`LSL` with `And` is performed by `DAGCombiner`.

However, in certain cases, the `AND` generated by this process
can be removed.

Consider following case:
```
        lsr x8, x8, #56
        and x8, x8, #0xfc
        ldr w0, [x2, x8]
        ret
```

In this case, we can remove the `AND` by changing the target of `LDR`
to `[X2, X8, LSL #2]` and right-shifting amount change to 56 to 58.

after changed:
```
        lsr x8, x8, #58
        ldr w0, [x2, x8, lsl #2]
        ret
```

This patch checks to see if the `SHIFTING` + `AND` operation on load
target can be optimized and optimizes it if it can.

Test Plan: sandcastle

Reviewers: wangwei, h2h, #compiler_toolchain

Reviewed By: h2h

Differential Revision: https://phabricator.intern.facebook.com/D70356861

Tasks: T216562623

Tags: aarch64